# gateway/docs/swagger.yaml
basePath: /
...
host: localhost:8080
info:
  contact:
    email: support@erp-system.com
    name: ERP System Support
  description: Enterprise Resource Planning System API Gateway. Provides unified REST
    API access to Inventory, Order, Finance, and HR microservices.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: ERP System API
  version: "1.0"
paths:
...
  /api/v1/finance/accounts:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List all accounts
      tags:
      - Finance - Accounts
    post:
      consumes:
      - application/json
      description: Create a new account in the chart of accounts
      parameters:
      - description: Account data (code, name, type)
        in: body
        name: account
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create a chart of accounts entry
      tags:
      - Finance - Accounts
  /api/v1/finance/accounts/{id}:
    get:
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get an account by ID
      tags:
      - Finance - Accounts
  /api/v1/finance/budgets:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List all budgets
      tags:
      - Finance - Budgets
    post:
      consumes:
      - application/json
      parameters:
      - description: Budget data
        in: body
        name: budget
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create a budget
      tags:
      - Finance - Budgets
  /api/v1/finance/budgets/{id}:
    get:
      parameters:
      - description: Budget ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get a budget by ID
      tags:
      - Finance - Budgets
  /api/v1/finance/invoices:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List all invoices
      tags:
      - Finance - Invoices
    post:
      consumes:
      - application/json
      parameters:
      - description: Invoice data
        in: body
        name: invoice
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create an invoice
      tags:
      - Finance - Invoices
  /api/v1/finance/invoices/{id}:
    get:
      parameters:
      - description: Invoice ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get an invoice by ID
      tags:
      - Finance - Invoices
  /api/v1/finance/invoices/{id}/payment:
    post:
      consumes:
      - application/json
      parameters:
      - description: Invoice ID
        in: path
        name: id
        required: true
        type: string
      - description: Payment data
        in: body
        name: payment
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Record a payment against an invoice
      tags:
      - Finance - Invoices
  /api/v1/finance/journal-entries:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List all journal entries
      tags:
      - Finance - Journal Entries
    post:
      consumes:
      - application/json
      parameters:
      - description: Journal entry data (date, lines, memo)
        in: body
        name: entry
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create a journal entry
      tags:
      - Finance - Journal Entries
  /api/v1/finance/journal-entries/{id}:
    get:
      parameters:
      - description: Journal Entry ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get a journal entry by ID
      tags:
      - Finance - Journal Entries
  /api/v1/finance/journal-entries/{id}/post:
    post:
      description: Post a draft journal entry to the general ledger
      parameters:
      - description: Journal Entry ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Post a journal entry
      tags:
      - Finance - Journal Entries
  /api/v1/finance/journal-entries/{id}/reverse:
    post:
      parameters:
      - description: Journal Entry ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Reverse a journal entry
      tags:
      - Finance - Journal Entries
  /api/v1/finance/reports/ar-aging:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get accounts receivable aging report
      tags:
      - Finance - Reports
  /api/v1/finance/reports/balance-sheet:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get balance sheet report
      tags:
      - Finance - Reports
  /api/v1/finance/reports/profit-loss:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get profit & loss report
      tags:
      - Finance - Reports
  /api/v1/finance/reports/trial-balance:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get trial balance report
      tags:
      - Finance - Reports
  /api/v1/finance/tax/calculate:
    post:
      consumes:
      - application/json
      parameters:
      - description: Tax calculation input
        in: body
        name: request
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Calculate tax
      tags:
      - Finance - Tax
...
securityDefinitions:
  BearerAuth:
    description: Enter your JWT token with the Bearer prefix, e.g. "Bearer eyJhbGciOiJI..."
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...