# gateway/handlers/finance.go
package handlers

import (
	"net/http"

	"github.com/gin-gonic/gin"
	"google.golang.org/grpc"
)

type FinanceHandler struct {
	conn *grpc.ClientConn
}

func NewFinanceHandler(conn *grpc.ClientConn) *FinanceHandler {
	return &FinanceHandler{conn: conn}
}

// CreateAccount godoc
// @Summary      Create a chart of accounts entry
// @Description  Create a new account in the chart of accounts
// @Tags         Finance - Accounts
// @Accept       json
// @Produce      json
// @Param        account  body      map[string]interface{}  true  "Account data (code, name, type)"
// @Success      201      {object}  map[string]interface{}
// @Failure      400      {object}  map[string]interface{}
// @Security     BearerAuth
// @Router       /api/v1/finance/accounts [post]
func (h *FinanceHandler) CreateAccount(c *gin.Context) {
	var req map[string]interface{}
	if err := c.ShouldBindJSON(&req); err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		return
	}
	c.JSON(http.StatusCreated, gin.H{"message": "Account created", "data": req})
}

// GetAccount godoc
// @Summary      Get an account by ID
// @Tags         Finance - Accounts
// @Produce      json
// @Param        id   path      string  true  "Account ID"
// @Success      200  {object}  map[string]interface{}
// @Security     BearerAuth
// @Router       /api/v1/finance/accounts/{id} [get]
func (h *FinanceHandler) GetAccount(c *gin.Context) { c.JSON(http.StatusOK, gin.H{"id": c.Param("id")}) }

// ListAccounts godoc
// @Summary      List all accounts
// @Tags         Finance - Accounts
// @Produce      json
// @Success      200  {object}  map[string]interface{}
// @Security     BearerAuth
// @Router       /api/v1/finance/accounts [get]
func (h *FinanceHandler) ListAccounts(c *gin.Context) { c.JSON(http.StatusOK, gin.H{"message": "List accounts"}) }

// CreateJournalEntry godoc
// @Summary      Create a journal entry
// @Tags         Finance - Journal Entries
// @Accept       json
// @Produce      json
// @Param        entry  body      map[string]interface{}  true  "Journal entry data (date, lines, memo)"
// @Success      201    {object}  map[string]interface{}
// @Failure      400    {object}  map[string]interface{}
// @Security     BearerAuth
// @Router       /api/v1/finance/journal-entries [post]
func (h *FinanceHandler) CreateJournalEntry(c *gin.Context) {
	var req map[string]interface{}
	if err := c.ShouldBindJSON(&req); err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		return
	}
	c.JSON(http.StatusCreated, gin.H{"message": "Journal entry created", "data": req})
}

// GetJournalEntry godoc
// @Summary      Get a journal entry by ID
// @Tags         Finance - Journal Entries
// @Produce      json
// @Param        id   path      string  true  "Journal Entry ID"
// @Success      200  {object}  map[string]interface{}
// @Security     BearerAuth
// @Router       /api/v1/finance/journal-entries/{id} [get]
func (h *FinanceHandler) GetJournalEntry(c *gin.Context) { c.JSON(http.StatusOK, gin.H{"id": c.Param("id")}) }

// ListJournalEntries godoc
// @Summary      List all journal entries
// @Tags         Finance - Journal Entries
// @Produce      json
// @Success      200  {object}  map[string]interface{}
// @Security     BearerAuth
// @Router       /api/v1/finance/journal-entries [get]
func (h *FinanceHandler) ListJournalEntries(c *gin.Context) { c.JSON(http.StatusOK, gin.H{"message": "List entries"}) }

// PostJournalEntry godoc
// @Summary      Post a journal entry
// @Description  Post a draft journal entry to the general ledger
// @Tags         Finance - Journal Entries
// @Produce      json
// @Param        id   path      string  true  "Journal Entry ID"
// @Success      200  {object}  map[string]interface{}
// @Security     BearerAuth
// @Router       /api/v1/finance/journal-entries/{id}/post [post]
func (h *FinanceHandler) PostJournalEntry(c *gin.Context) {
	c.JSON(http.StatusOK, gin.H{"message": "Journal entry posted", "id": c.Param("id")})
}

// ReverseJournalEntry godoc
// @Summary      Reverse a journal entry
// @Tags         Finance - Journal Entries
// @Produce      json
// @Param        id   path      string  true  "Journal Entry ID"
// @Success      200  {object}  map[string]interface{}
// @Security     BearerAuth
// @Router       /api/v1/finance/journal-entries/{id}/reverse [post]
func (h *FinanceHandler) ReverseJournalEntry(c *gin.Context) {
	c.JSON(http.StatusOK, gin.H{"message": "Journal entry reversed", "id": c.Param("id")})
}

// CreateInvoice godoc
// @Summary      Create an invoice
// @Tags         Finance - Invoices
// @Accept       json
// @Produce      json
// @Param        invoice  body      map[string]interface{}  true  "Invoice data"
// @Success      201      {object}  map[string]interface{}
// @Failure      400      {object}  map[string]interface{}
// @Security     BearerAuth
// @Router       /api/v1/finance/invoices [post]
func (h *FinanceHandler) CreateInvoice(c *gin.Context) {
	var req map[string]interface{}
	if err := c.ShouldBindJSON(&req); err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		return
	}
	c.JSON(http.StatusCreated, gin.H{"message": "Invoice created", "data": req})
}

// GetInvoice godoc
// @Summary      Get an invoice by ID
// @Tags         Finance - Invoices
// @Produce      json
// @Param        id   path      string  true  "Invoice ID"
// @Success      200  {object}  map[string]interface{}
// @Security     BearerAuth
// @Router       /api/v1/finance/invoices/{id} [get]
func (h *FinanceHandler) GetInvoice(c *gin.Context) { c.JSON(http.StatusOK, gin.H{"id": c.Param("id")}) }

// ListInvoices godoc
// @Summary      List all invoices
// @Tags         Finance - Invoices
// @Produce      json
// @Success      200  {object}  map[string]interface{}
// @Security     BearerAuth
// @Router       /api/v1/finance/invoices [get]
func (h *FinanceHandler) ListInvoices(c *gin.Context) { c.JSON(http.StatusOK, gin.H{"message": "List invoices"}) }

// RecordPayment godoc
// @Summary      Record a payment against an invoice
// @Tags         Finance - Invoices
// @Accept       json
// @Produce      json
// @Param        id       path      string                 true  "Invoice ID"
// @Param        payment  body      map[string]interface{}  true  "Payment data"
// @Success      200      {object}  map[string]interface{}
// @Failure      400      {object}  map[string]interface{}
// @Security     BearerAuth
// @Router       /api/v1/finance/invoices/{id}/payment [post]
func (h *FinanceHandler) RecordPayment(c *gin.Context) {
	var req map[string]interface{}
	if err := c.ShouldBindJSON(&req); err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		return
	}
	c.JSON(http.StatusOK, gin.H{"message": "Payment recorded", "invoice_id": c.Param("id"), "data": req})
}

// CalculateTax godoc
// @Summary      Calculate tax
// @Tags         Finance - Tax
// @Accept       json
// @Produce      json
// @Param        request  body      map[string]interface{}  true  "Tax calculation input"
// @Success      200      {object}  map[string]interface{}
// @Failure      400      {object}  map[string]interface{}
// @Security     BearerAuth
// @Router       /api/v1/finance/tax/calculate [post]
func (h *FinanceHandler) CalculateTax(c *gin.Context) {
	var req map[string]interface{}
	if err := c.ShouldBindJSON(&req); err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		return
	}
	c.JSON(http.StatusOK, gin.H{"message": "Tax calculated", "data": req})
}

// GetTrialBalance godoc
// @Summary      Get trial balance report
// @Tags         Finance - Reports
// @Produce      json
// @Success      200  {object}  map[string]interface{}
// @Security     BearerAuth
// @Router       /api/v1/finance/reports/trial-balance [get]
func (h *FinanceHandler) GetTrialBalance(c *gin.Context) { c.JSON(http.StatusOK, gin.H{"report": "trial_balance"}) }

// GetProfitAndLoss godoc
// @Summary      Get profit & loss report
// @Tags         Finance - Reports
// @Produce      json
// @Success      200  {object}  map[string]interface{}
// @Security     BearerAuth
// @Router       /api/v1/finance/reports/profit-loss [get]
func (h *FinanceHandler) GetProfitAndLoss(c *gin.Context) { c.JSON(http.StatusOK, gin.H{"report": "profit_loss"}) }

// GetBalanceSheet godoc
// @Summary      Get balance sheet report
// @Tags         Finance - Reports
// @Produce      json
// @Success      200  {object}  map[string]interface{}
// @Security     BearerAuth
// @Router       /api/v1/finance/reports/balance-sheet [get]
func (h *FinanceHandler) GetBalanceSheet(c *gin.Context) { c.JSON(http.StatusOK, gin.H{"report": "balance_sheet"}) }

// GetARaging godoc
// @Summary      Get accounts receivable aging report
// @Tags         Finance - Reports
// @Produce      json
// @Success      200  {object}  map[string]interface{}
// @Security     BearerAuth
// @Router       /api/v1/finance/reports/ar-aging [get]
func (h *FinanceHandler) GetARaging(c *gin.Context) { c.JSON(http.StatusOK, gin.H{"report": "ar_aging"}) }

// CreateBudget godoc
// @Summary      Create a budget
// @Tags         Finance - Budgets
// @Accept       json
// @Produce      json
// @Param        budget  body      map[string]interface{}  true  "Budget data"
// @Success      201     {object}  map[string]interface{}
// @Failure      400     {object}  map[string]interface{}
// @Security     BearerAuth
// @Router       /api/v1/finance/budgets [post]
func (h *FinanceHandler) CreateBudget(c *gin.Context) {
	var req map[string]interface{}
	if err := c.ShouldBindJSON(&req); err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		return
	}
	c.JSON(http.StatusCreated, gin.H{"message": "Budget created", "data": req})
}

// GetBudget godoc
// @Summary      Get a budget by ID
// @Tags         Finance - Budgets
// @Produce      json
// @Param        id   path      string  true  "Budget ID"
// @Success      200  {object}  map[string]interface{}
// @Security     BearerAuth
// @Router       /api/v1/finance/budgets/{id} [get]
func (h *FinanceHandler) GetBudget(c *gin.Context) { c.JSON(http.StatusOK, gin.H{"id": c.Param("id")}) }

// ListBudgets godoc
// @Summary      List all budgets
// @Tags         Finance - Budgets
// @Produce      json
// @Success      200  {object}  map[string]interface{}
// @Security     BearerAuth
// @Router       /api/v1/finance/budgets [get]
func (h *FinanceHandler) ListBudgets(c *gin.Context) { c.JSON(http.StatusOK, gin.H{"message": "List budgets"}) }